doe {radiant.design}	R Documentation

Create (partial) factorial design

Description

Create (partial) factorial design

Usage

doe(factors, int = "", trials = NA, seed = NA)

Arguments

factors
Categorical variables used as input for design

int
Vector of interaction terms to consider when generating design

trials
Number of trials to create. If NA then all feasible designs will be considered until a design with perfect D-efficiency is found

seed
Random seed to use as the starting point

Details

See https://radiant-rstats.github.io/docs/design/doe.html for an example in Radiant

Value

A list with all variables defined in the function as an object of class doe

See Also

summary.doe to summarize results

Examples

doe(c("price; $10; $13; $16", "food; popcorn; gourmet; no food"))
doe(
  c("price; $10; $13; $16", "food; popcorn; gourmet; no food"),
  int = "price:food", trials = 9, seed = 1234
)

[Package radiant.design version 0.9.9 Index]